pyproject.toml
[project]
name = "seliseblocks-lmt"
version = "0.0.3"
readme = "README.md"
description = "Logging, Monitoring, and Tracing library for Selise Blocks applications using Azure Service Bus."
authors = [{name="Selise Blocks", email="mostafizur.rahman@selisegroup.com"}]
keywords = ["Selise", "Selise Blocks", "LMT", "Logging", "Monitoring", "Tracing", "Azure Service Bus", "OpenTelemetry"]
dependencies = [
    "azure-servicebus>=7.14.2",
    "opentelemetry-api>=1.33.1",
    "opentelemetry-sdk>=1.33.1",
    "opentelemetry-instrumentation-fastapi>=0.54b1",
]

requires-python = ">=3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["blocks_lmt"]

# Optional dependencies for development and testing
[project.optional-dependencies]
# For running tests
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "httpx>=0.28.0",
]

# For development (includes test + extra tools)
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "httpx>=0.28.0",
    "debugpy>=1.8.0",
    "fastapi>=0.126.0",
    "uvicorn>=0.40.0",
    "pytest-cov>=4.1.0",
]

[project.urls]
Homepage = "https://github.com/SELISEdigitalplatforms/l0-py-blocks-genesis"
Repository = "https://github.com/SELISEdigitalplatforms/l0-py-blocks-genesis"